task file upload-image
unload.html

+++ files/app.py
@@ -1,6 +1,7 @@
-from fastapi import FastAPI, UploadFile, File,HTTPException
+from fastapi import FastAPI, UploadFile, File, HTTPException
 from fastapi.responses import HTMLResponse
 from fastapi.staticfiles import StaticFiles
+
 import os
 import shutil
 
@@ -19,6 +20,7 @@ async def read_root():
 
 MAX_FILE_SIZE = 5 * 1024 * 1024  # 5 MB
 ALLOWED_FILE_TYPES = ["image/png", "image/jpeg", "image/gif"]
+ALLOWED_FILE_TYPES_2 = ["application/pdf", "text/plain", "application/zip"]
 
 
 @app.post("/upload-image")
@@ -35,11 +37,19 @@ async def upload_image(file: UploadFile = File(...)):
         raise HTTPException(
             status_code=413,
             detail=f"File is too large. Max size allowed is {MAX_FILE_SIZE//(1024*1024)} MB."
+
+        )
+
+    file_path = os.path.join("static/temp", file.filename)
+    if os.path.exists(file_path):
+        raise HTTPException(
+            status_code=400,
+            detail="A file with the same name already exists. Please rename your file and try again."
         )
 
     file_path = f"static/temp/{file.filename}"
     with open(file_path, 'wb') as buffer:
-        shutil.copyfileobj(file.file, buffer)
+        buffer.write(file_content)
 
     return {"filename": file.filename, "file_path": file_path}
